requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
exceptiongroup==1.2.2
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.5
langchain-core==0.2.41
langchain-openai==0.1.17
langchain-teddynote==0.0.6
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.137
lxml==5.3.0
marshmallow==3.23.0
matplotlib==3.9.0
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.52.2
opencv-python==4.10.0.82
orjson==3.10.10
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.10.19   
XlsxWriter==3.2.0
yarl==1.17.0
